Final changes to directory structure. 'app' folder has been renamed to 'gui', 'time_slot_config_editor' folder has been renamed to 'timeslot_config', & a new 'ai' folder has been created.
