+++ ai/ai_assistant.py
