+++ gui/schedule_config_manager.py
